% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bent_arrows.R
\name{add_bent_arrows_xx}
\alias{add_bent_arrows_xx}
\title{Title}
\usage{
add_bent_arrows_xx(
  xy,
  M,
  mnwd = 0.05,
  bbend = 1,
  adj = 2,
  endd = 0.75,
  clr = "red"
)
}
\arguments{
\item{xy}{the point set}

\item{M}{the dispersal matrix}

\item{mnwd}{the minimum width to plot}

\item{bbend}{a parameter to bend arrows}

\item{adj}{set the maximum cex for points}

\item{endd}{relative position of arrowhead}

\item{clr}{the color}
}
\value{
invisible(NULL)
}
\description{
Title
}
