% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convex_hulls.R
\name{add_convex_hulls}
\alias{add_convex_hulls}
\title{Add convex hulls around the points in a community}
\usage{
add_convex_hulls(memix, xy, clrs, stretch = 0.1, lwd = 2, llty = 1)
}
\arguments{
\item{memix}{the membership index}

\item{xy}{the xy locasions}

\item{clrs}{colors}

\item{stretch}{a stretch factor for the hulls}

\item{lwd}{the line width for plotting}

\item{llty}{the line type}
}
\value{
invisible(NULL)
}
\description{
Add convex hulls around the points in a community
}
