% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_arrows.R, R/plot_arrows.R
\name{add_nth_arrow_xx}
\alias{add_nth_arrow_xx}
\title{Plot the n^th arrow}
\usage{
add_nth_arrow_xx(
  n,
  ij,
  xy,
  p = 1,
  r = 0,
  arw_lng = 0.1,
  wt = 1,
  lamp = 1,
  clr = "#CCCCCC"
)

add_nth_arrow_xx(
  n,
  ij,
  xy,
  p = 1,
  r = 0,
  arw_lng = 0.1,
  wt = 1,
  lamp = 1,
  clr = "#CCCCCC"
)
}
\arguments{
\item{n}{the arrow to plot}

\item{ij}{the indices of the arrows to plot}

\item{xy}{the xy locations}

\item{p}{fraction of length to plot}

\item{r}{the radius of a ring around destination points}

\item{arw_lng}{the arrow length}

\item{wt}{the line weight}

\item{lamp}{arrow width scaling factor}

\item{clr}{the arrow color}
}
\value{
invisible(NULL)

invisible(NULL)
}
\description{
Plot the n^th arrow

Plot the n^th arrow
}
