% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_arrows.R, R/plot_arrows.R
\name{add_nth_arrow_xy}
\alias{add_nth_arrow_xy}
\title{Add the nth arrow from a list of locations and indices}
\usage{
add_nth_arrow_xy(
  n,
  ij,
  xy_launch,
  xy_land,
  p = 1,
  r = 0,
  arw_lng = 0.1,
  wt = 1,
  lamp = 1,
  clr = "#CCCCCC"
)

add_nth_arrow_xy(
  n,
  ij,
  xy_launch,
  xy_land,
  p = 1,
  r = 0,
  arw_lng = 0.1,
  wt = 1,
  lamp = 1,
  clr = "#CCCCCC"
)
}
\arguments{
\item{n}{the arrow to plot}

\item{ij}{the indices of the arrows to plot}

\item{xy_launch}{the xy location of the search origin}

\item{xy_land}{the xy location of the search end}

\item{p}{fraction of length to plot}

\item{r}{the radius of a ring around destination points}

\item{arw_lng}{the arrow length}

\item{wt}{the line weight}

\item{lamp}{arrow width scaling factor}

\item{clr}{the arrow color}
}
\value{
invisible(NULL)

invisible(NULL)
}
\description{
Add the nth arrow from a list of locations and indices

Add the nth arrow from a list of locations and indices
}
