% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_arrows.R, R/plot_arrows.R
\name{add_nth_segment}
\alias{add_nth_segment}
\title{Add the nth segment}
\usage{
add_nth_segment(n, ij, xy, wt = 1, lamp = 1, clr = "#CCCCCC")

add_nth_segment(n, ij, xy, wt = 1, lamp = 1, clr = "#CCCCCC")
}
\arguments{
\item{n}{an index for}

\item{ij}{the indices of the arrows to plot}

\item{xy}{the xy locations}

\item{wt}{the line weight}

\item{lamp}{arrow width scaling factor}

\item{clr}{the arrow color}
}
\value{
invisible(NULL)

invisible(NULL)
}
\description{
Add the nth segment

Add the nth segment
}
