% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_arrows.R, R/plot_arrows.R
\name{add_one_arrow}
\alias{add_one_arrow}
\title{Draw an arrow from (x_i, y_i) to (x_j, y_j)}
\usage{
add_one_arrow(
  x_i,
  y_i,
  x_j,
  y_j,
  p = 1,
  r = 0,
  arw_lng = 0.1,
  wt = 1,
  lamp = 1,
  clr = "#CCCCCC"
)

add_one_arrow(
  x_i,
  y_i,
  x_j,
  y_j,
  p = 1,
  r = 0,
  arw_lng = 0.1,
  wt = 1,
  lamp = 1,
  clr = "#CCCCCC"
)
}
\arguments{
\item{x_i}{source x value}

\item{y_i}{source y value}

\item{x_j}{destination x value}

\item{y_j}{destination y value}

\item{p}{fraction of length to plot}

\item{r}{the radius of a ring around destination points}

\item{arw_lng}{the arrow length}

\item{wt}{the line weight}

\item{lamp}{arrow width scaling factor}

\item{clr}{the arrow color}
}
\value{
invisible(NULL)

invisible(NULL)
}
\description{
Draw an arrow from (x_i, y_i) to (x_j, y_j)

Draw an arrow from (x_i, y_i) to (x_j, y_j)
}
