% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_points.R
\name{add_points_b}
\alias{add_points_b}
\title{Add blood feeding points to a frame}
\usage{
add_points_b(b, wts = 1, pw = 1, max_pt_sz = 1, clr = "#fe5f55CC")
}
\arguments{
\item{b}{a point set defining blood feeding sites}

\item{wts}{values to scale the point size}

\item{pw}{power relationship for scaling point size: pw=1 is linear}

\item{max_pt_sz}{set the maximum cex for points}

\item{clr}{the color}
}
\value{
invisible(NULL)
}
\description{
Add blood feeding points to a frame
}
