% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_points.R, R/viz_points.R
\name{add_points_bb}
\alias{add_points_bb}
\title{Add blood feeding sites with and without the self loop}
\usage{
add_points_bb(
  b,
  M,
  cx_b = 2,
  clr_bA = "#cc444b66",
  clr_bB = "#cc444bCC",
  pw = 1
)

add_points_bb(
  b,
  M,
  cx_b = 2,
  clr_bA = "#cc444b66",
  clr_bB = "#cc444bCC",
  pw = 1
)
}
\arguments{
\item{b}{a point set defining blood feeding sites}

\item{M}{a matrix describing movement to b}

\item{cx_b}{set the maximum cex for points}

\item{clr_bA}{color with the self loop}

\item{clr_bB}{color without the self loop}

\item{pw}{power relationship for scaling point size: pw=1 is linear}
}
\value{
invisible(NULL)

invisible(NULL)
}
\description{
Add blood feeding sites with and without the self loop

Add blood feeding sites with and without the self loop
}
