% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_points.R, R/viz_points.R
\name{add_points_qq}
\alias{add_points_qq}
\title{Add habitats with and without the self loop}
\usage{
add_points_qq(q, M, cx_q = 2, clr_qA = "skyblue", clr_qB = "skyblue3", pw = 1)

add_points_qq(q, M, cx_q = 2, clr_qA = "skyblue", clr_qB = "skyblue3", pw = 1)
}
\arguments{
\item{q}{a point set defining egg laying sites}

\item{M}{a matrix describing movement to q}

\item{cx_q}{set the maximum cex for points}

\item{clr_qA}{color with the self loop}

\item{clr_qB}{color without the self loop}

\item{pw}{power relationship for scaling point size: pw=1 is linear}
}
\value{
invisible(NULL)

invisible(NULL)
}
\description{
Add habitats with and without the self loop

Add habitats with and without the self loop
}
