% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_points.R, R/viz_points.R
\name{add_points_s}
\alias{add_points_s}
\title{Add sugar feeding sites to a frame}
\usage{
add_points_s(s, wts_s = 1, cx_s = 1, clr_s = "olivedrab2", pw = 1)

add_points_s(s, wts_s = 1, cx_s = 1, clr_s = "olivedrab2", pw = 1)
}
\arguments{
\item{s}{a point set defining sugar feeding sites}

\item{wts_s}{values to scale the point size}

\item{cx_s}{set the maximum cex for points}

\item{clr_s}{the color}

\item{pw}{power relationship for scaling point size: pw=1 is linear}
}
\value{
invisible(NULL)

invisible(NULL)
}
\description{
Add sugar feeding sites to a frame

Add sugar feeding sites to a frame
}
