% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_points.R
\name{add_points_ss}
\alias{add_points_ss}
\title{Add sugar feeding sites with and without the self loop}
\usage{
add_points_ss(
  s,
  M,
  cx_s = 2,
  clr_sA = "olivedrab2",
  clr_sB = "olivedrab3",
  pw = 1
)
}
\arguments{
\item{s}{a point set defining sugar feeding sites}

\item{M}{a matrix describing movement to s}

\item{cx_s}{set the maximum cex for points}

\item{clr_sA}{color with the self loop}

\item{clr_sB}{color without the self loop}

\item{pw}{power relationship for scaling point size: pw=1 is linear}
}
\value{
invisible(NULL)
}
\description{
Add sugar feeding sites with and without the self loop
}
