% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-BQ.R
\name{adult_dynamics.BQ}
\alias{adult_dynamics.BQ}
\title{Simulate adult dynamics for the `BQ` model}
\usage{
\method{adult_dynamics}{BQ}(t, model)
}
\arguments{
\item{t}{the current time}

\item{model}{a model defined as a compound [list]}
}
\value{
model, a compound [list]
}
\description{
Simulate adult dynamics for the `BQ` model
}
