% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-interface.R
\name{adult_dynamics}
\alias{adult_dynamics}
\title{Adult dynamics}
\usage{
adult_dynamics(t, model)
}
\arguments{
\item{t}{the current time}

\item{model}{a model defined as a compound [list]}
}
\value{
the model, a compound [list]
}
\description{
Adult dynamics
}
