% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-BQS.R
\name{AdultDynamics.BQS}
\alias{AdultDynamics.BQS}
\title{Simulate adult dynamics for the `BQ` model: one time step}
\usage{
\method{AdultDynamics}{BQS}(t, model)
}
\arguments{
\item{t}{the current time}

\item{model}{a model defined as a compound [list]}
}
\value{
model, a compound [list]
}
\description{
Simulate adult dynamics for the `BQ` model: one time step
}
