% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-interface.R
\name{aquatic_dynamics}
\alias{aquatic_dynamics}
\title{Aquatic Dynamics}
\usage{
aquatic_dynamics(t, model)
}
\arguments{
\item{t}{the current time}

\item{model}{a model defined as a compound [list]}
}
\value{
the model, a compound [list]
}
\description{
Aquatic Dynamics
}
