% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-basicL.R
\name{AquaticDynamics.basicL}
\alias{AquaticDynamics.basicL}
\title{Aquatic Dynamics}
\usage{
\method{AquaticDynamics}{basicL}(t, model)
}
\arguments{
\item{t}{the current time}

\item{model}{a model defined as a compound [list]}
}
\value{
the model, a compound [list]
}
\description{
Aquatic Dynamics
}
