% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point_sets.R
\name{clusters_nm}
\alias{clusters_nm}
\title{Setup n clusters with m points per cluster}
\usage{
clusters_nm(n, m, vr, mn, mx)
}
\arguments{
\item{n}{the number of seed points points}

\item{m}{the number of points per cluster}

\item{vr}{the variance in the dispersion around the seed}

\item{mn}{the lower limit for location values}

\item{mx}{the upper limit for location values}
}
\value{
a set of xy locations
}
\description{
Setup n clusters with m points per cluster
}
