% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point_sets.R
\name{clusters_xy}
\alias{clusters_xy}
\title{Use a set of `xy` values as seeds for clusters}
\usage{
clusters_xy(xy, nc = 1, vr = 1)
}
\arguments{
\item{xy}{a set of seeds}

\item{nc}{the number of points per cluster}

\item{vr}{the variance in the dispersion around the seed}
}
\value{
a set of xy locations
}
\description{
Use a set of `xy` values as seeds for clusters
}
