% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-interface.R
\name{compute_diffs_L}
\alias{compute_diffs_L}
\title{The some of squared differences between two sets of variables}
\usage{
compute_diffs_L(Lvars1, Lvars2)
}
\arguments{
\item{Lvars1}{variables describing adult mosquitoes, set 1}

\item{Lvars2}{variables describing adult mosquitoes, set 2}
}
\value{
a [numeric] value, the sum of squared differences
}
\description{
The some of squared differences between two sets of variables
}
