% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-BQ.R, R/adult-BQS.R
\name{compute_diffs_M.BQ}
\alias{compute_diffs_M.BQ}
\title{The some of squared differences between two sets of variables}
\usage{
\method{compute_diffs_M}{BQ}(Mvars1, Mvars2)

\method{compute_diffs_M}{BQ}(Mvars1, Mvars2)
}
\arguments{
\item{Mvars1}{variables describing adult mosquitoes, set 1}

\item{Mvars2}{variables describing adult mosquitoes, set 2}
}
\value{
a [numeric] value, the sum of squared differences

a [numeric] value, the sum of squared differences
}
\description{
The some of squared differences between two sets of variables

The some of squared differences between two sets of variables
}
