% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-BQS.R
\name{compute_diffs_M.BQS}
\alias{compute_diffs_M.BQS}
\title{The sum of squared differences between the state and putative steady state}
\usage{
\method{compute_diffs_M}{BQS}(model)
}
\arguments{
\item{model}{a model}
}
\value{
a [numeric] value, the sum of squared differences
}
\description{
The sum of squared differences between the state and putative steady state
}
