% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-interface.R
\name{compute_diffs_M}
\alias{compute_diffs_M}
\title{The some of squared differences between two sets of variables}
\usage{
compute_diffs_M(model)
}
\arguments{
\item{model}{a model}
}
\value{
a [numeric] value, the sum of squared differences
}
\description{
The some of squared differences between two sets of variables
}
