% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_G.R
\name{compute_G.BQS}
\alias{compute_G.BQS}
\title{Compute egg dispersion, BQS model}
\usage{
\method{compute_G}{BQS}(model, Tmax = 50)
}
\arguments{
\item{model}{a compound [list] defining a model}

\item{Tmax}{the last time step}
}
\value{
the model, a compound [list]
}
\description{
Compute egg dispersion, BQS model
}
