% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_G.R
\name{compute_G}
\alias{compute_G}
\title{Compute egg dispersion}
\usage{
compute_G(model, Tmax)
}
\arguments{
\item{model}{a compound [list] defining a model}

\item{Tmax}{the last time step}
}
\value{
the model, a compound [list]
}
\description{
Compute the dispersion of eggs away from natal habitats.
}
