% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_K.R
\name{compute_Kbb}
\alias{compute_Kbb}
\title{Compute dispersal matrix to lay eggs within one feeding cycle}
\usage{
compute_Kbb(model)
}
\arguments{
\item{model}{a model defined as a compound [list]}
}
\value{
the model, a compound [list]
}
\description{
Compute dispersal matrix to lay eggs within one feeding cycle
}
