% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_model.R
\name{compute_KGV}
\alias{compute_KGV}
\title{Compute net dispersal matrix to blood feed within one feeding cycle}
\usage{
compute_KGV(model, Tmax = 200)
}
\arguments{
\item{model}{a model defined as a compound [list]}

\item{Tmax}{the number of time steps}
}
\value{
the model, a compound [list]
}
\description{
Compute net dispersal matrix to blood feed within one feeding cycle
}
