% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_K.R
\name{compute_Kqb.BQ}
\alias{compute_Kqb.BQ}
\title{Compute net dispersal matrix to blood feed within one feeding cycle}
\usage{
\method{compute_Kqb}{BQ}(model, Tmax = 100)
}
\arguments{
\item{model}{a model defined as a compound [list]}

\item{Tmax}{the number of time steps}
}
\value{
the model, a compound [list]
}
\description{
Compute net dispersal matrix to blood feed within one feeding cycle
}
