% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_K.R
\name{compute_Kqb}
\alias{compute_Kqb}
\title{Compute dispersal matrix to lay eggs within one feeding cycle}
\usage{
compute_Kqb(model, Tmax)
}
\arguments{
\item{model}{a model defined as a compound [list]}

\item{Tmax}{the number of time steps}
}
\value{
the model, a compound [list]
}
\description{
Compute dispersal matrix to lay eggs within one feeding cycle
}
