% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_V.R
\name{compute_V.BQ}
\alias{compute_V.BQ}
\title{Compute the potential dispersion of parasites after one feeding cycle for the BQ model}
\usage{
\method{compute_V}{BQ}(model, Tmax = 100)
}
\arguments{
\item{model}{a compound [list] defining a model}

\item{Tmax}{the last time step}
}
\value{
the model, a compound [list]
}
\description{
Compute the potential dispersion of parasites after one feeding cycle for the BQ model
}
