% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_V.R
\name{compute_V}
\alias{compute_V}
\title{Compute the potential dispersion of parasites after one feeding cycle}
\usage{
compute_V(model, Tmax)
}
\arguments{
\item{model}{a compound [list] defining a model}

\item{Tmax}{the last time step}
}
\value{
the model, a compound [list]
}
\description{
Compute the potential dispersion of parasites after one feeding cycle
}
