% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_V.R
\name{compute_VC}
\alias{compute_VC}
\title{Compute the potential dispersion of parasites after one feeding cycle for the BQS model}
\usage{
compute_VC(model)
}
\arguments{
\item{model}{a compound [list] defining a model}
}
\value{
the model, a compound [list]
}
\description{
Compute the potential dispersion of parasites after one feeding cycle for the BQS model
}
