% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egg_laying_compute.R
\name{computeG}
\alias{computeG}
\title{Compute the dispersion of eggs after one feeding cycle}
\usage{
computeG(model, Tmax)
}
\arguments{
\item{model}{a compound [list] defining a model}

\item{Tmax}{the last time step}
}
\value{
the model, a compound [list]
}
\description{
Compute the dispersion of eggs after one feeding cycle
}
