% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egg_laying_compute.R
\name{computeGG}
\alias{computeGG}
\title{Compute the dispersion of eggs after one feeding cycle for the BQ model}
\usage{
computeGG(model)
}
\arguments{
\item{model}{a compound [list] defining a model}
}
\value{
the model, a compound [list]
}
\description{
Compute the dispersion of eggs after one feeding cycle for the BQ model
}
