% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectorial_capacity_compute.R
\name{computeV}
\alias{computeV}
\title{Compute the potential dispersion of parasites after one feeding cycle}
\usage{
computeV(model, Tmax)
}
\arguments{
\item{model}{a compound [list] defining a model}

\item{Tmax}{the last time step}
}
\value{
the model, a compound [list]
}
\description{
Compute the potential dispersion of parasites after one feeding cycle
}
