% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectorial_capacity_compute.R
\name{computeVC}
\alias{computeVC}
\title{Compute the potential dispersion of parasites after one feeding cycle for the BQS model}
\usage{
computeVC(model)
}
\arguments{
\item{model}{a compound [list] defining a model}
}
\value{
the model, a compound [list]
}
\description{
Compute the potential dispersion of parasites after one feeding cycle for the BQS model
}
