% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cutoffValue}
\alias{cutoffValue}
\title{Define a cutoff that includes a fraction of the total mass (e.g. 90 percent)}
\usage{
cutoffValue(M, fracMass = 0.9)
}
\arguments{
\item{M}{a matrix}

\item{fracMass}{the fraction of mass to plot}
}
\value{
a [numeric] value
}
\description{
Define a cutoff that includes a fraction of the total mass (e.g. 90 percent)
}
