% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{decompM}
\alias{decompM}
\title{Decompose Matrix}
\usage{
decompM(M)
}
\arguments{
\item{M}{a matrix}
}
\value{
a [list] of the 3 decomposed matrices
}
\description{
Decompose a matrix into three parts: the
diagonal (self) loop, a symmetric matrix, and a flow
}
