% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{edgeSubset}
\alias{edgeSubset}
\title{Subset Edges}
\usage{
edgeSubset(M, min_edge_frac = 0.01)
}
\arguments{
\item{M}{a matrix}

\item{min_edge_frac}{the minimum fraction to plot edge}
}
\value{
a set of array indices
}
\description{
Finds the indices of the edges that comprise a given
fraction of the total dispersal mass
}
