% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{edgeSubset_fracMass}
\alias{edgeSubset_fracMass}
\title{Subset Edges}
\usage{
edgeSubset_fracMass(M, fracMass = 0.9)
}
\arguments{
\item{M}{a matrix}

\item{fracMass}{the fraction of mass to plot}
}
\value{
a set of array indices
}
\description{
Finds the indices of the edges that comprise a given
fraction of the total dispersal mass
}
