% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_points.R
\name{frame_bq}
\alias{frame_bq}
\title{Set up a frame to plot points for a model}
\usage{
frame_bq(b, q, mtl = "")
}
\arguments{
\item{b}{a point set defining blood feeding sites}

\item{q}{a point set defining egg laying sites}

\item{mtl}{a plot title}
}
\value{
invisible(NULL)
}
\description{
Set up a frame to plot points for a model
}
