% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_points.R, R/viz_points.R
\name{frame_bqs}
\alias{frame_bqs}
\title{Set up a frame to plot points for a BQS model}
\usage{
frame_bqs(b, q, s, mtl = "")

frame_bqs(b, q, s, mtl = "")
}
\arguments{
\item{b}{a point set defining blood feeding sites}

\item{q}{a point set defining egg laying sites}

\item{s}{a point set defining sugar feeding sites}

\item{mtl}{a plot title}
}
\value{
invisible(NULL)

invisible(NULL)
}
\description{
Set up a frame to plot points for a BQS model

Set up a frame to plot points for a BQS model
}
