% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{get_graph}
\alias{get_graph}
\title{Choose one of the networks describing dispersion}
\usage{
get_graph(model, i)
}
\arguments{
\item{model}{a model defined as a compound [list]}

\item{i}{an index for the type of model}
}
\value{
a net object
}
\description{
1 -Kbb; 2 - Kqq;
3 - G; 4 - GG;
5 - V; 6 - VV;
7 - M; 8 - MM
}
