% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_matrix}
\alias{get_matrix}
\title{Choose one of the matrices describing movement}
\usage{
get_matrix(i, model)
}
\arguments{
\item{i}{an index for the type of model}

\item{model}{a model defined as a compound [list]}
}
\value{
a dispersion matrix
}
\description{
1. Kbb; 2. Kqq;
3. G; 4. GG;
5. V; 6. VV;
7. M; 8. MM;
}
