% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getM_i}
\alias{getM_i}
\title{Choose one of the matrices describing movement}
\usage{
getM_i(i, model)
}
\arguments{
\item{i}{an index for the type of model}

\item{model}{a model defined as a compound [list]}
}
\value{
a dispersion matrix
}
\description{
Choose one of the matrices describing movement
}
