% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getNet_i}
\alias{getNet_i}
\title{Choose one of the networks describing dispersion}
\usage{
getNet_i(model, i)
}
\arguments{
\item{model}{a model defined as a compound [list]}

\item{i}{an index for the type of model}
}
\value{
a net object
}
\description{
Choose one of the networks describing dispersion
}
