% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-BQ.R
\name{init_adult_model.BQ}
\alias{init_adult_model.BQ}
\title{Set initial values for the BQ model}
\usage{
\method{init_adult_model}{BQ}(model, M0_opts)
}
\arguments{
\item{model}{a model defined as a compound [list]}

\item{M0_opts}{a list of options to overwrite defaults}
}
\value{
model, a compound [list]
}
\description{
Set initial values for the BQ model
}
