% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-BQS.R
\name{init_adult_model.BQS}
\alias{init_adult_model.BQS}
\title{Set initial values for the BQS model}
\usage{
\method{init_adult_model}{BQS}(model, M0_opts)
}
\arguments{
\item{model}{a model defined as a compound [list]}

\item{M0_opts}{a list of options to overwrite defaults}
}
\value{
model, a compound [list]
}
\description{
Set initial values for the BQS model
}
