% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-interface.R
\name{init_adult_model}
\alias{init_adult_model}
\title{Set Initial Values for **M** Module}
\usage{
init_adult_model(model, M0_opts)
}
\arguments{
\item{model}{a model defined as a compound [list]}

\item{M0_opts}{a list of options to overwrite defaults}
}
\value{
the model, a compound [list]
}
\description{
Set Initial Values for **M** Module
}
