% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-BQ.R
\name{init_adult_model_BQ}
\alias{init_adult_model_BQ}
\title{Set initial values for the BQ model}
\usage{
init_adult_model_BQ(model, opts = list(), B0 = 10, Q0 = 10)
}
\arguments{
\item{model}{a model defined as a compound [list]}

\item{opts}{a list of options to overwrite defaults}

\item{B0}{initial values for blood feeding mosquito density}

\item{Q0}{initial values for egg laying mosquito density}
}
\value{
model, a compound [list]
}
\description{
Set initial values for the BQ model
}
