% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-basicL.R
\name{init_aquatic_model.basicL}
\alias{init_aquatic_model.basicL}
\title{Set initial values for the BQ model}
\usage{
\method{init_aquatic_model}{basicL}(model, L0_opts)
}
\arguments{
\item{model}{a model defined as a compound [list]}

\item{L0_opts}{a list of options to overwrite defaults}
}
\value{
model, a compound [list]
}
\description{
Set initial values for the BQ model
}
