% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-interface.R
\name{init_aquatic_model}
\alias{init_aquatic_model}
\title{Set initial values for}
\usage{
init_aquatic_model(model, L0_opts = list())
}
\arguments{
\item{model}{a model defined as a compound [list]}

\item{L0_opts}{a list of options to overwrite defaults}
}
\value{
the model, a compound [list]
}
\description{
Set initial values for
}
