% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-basicL.R
\name{init_aquatic_model_basicL}
\alias{init_aquatic_model_basicL}
\title{Set initial values for}
\usage{
init_aquatic_model_basicL(model, opts = list(), L0 = 10)
}
\arguments{
\item{model}{a model defined as a compound [list]}

\item{opts}{a list of options to overwrite defaults}

\item{L0}{default initial value for larval density}
}
\value{
the model, a compound [list]
}
\description{
Set initial values for
}
