% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-basicL.R
\name{init_states_L.basicL}
\alias{init_states_L.basicL}
\title{Save the state variables in a vector}
\usage{
\method{init_states_L}{basicL}(model)
}
\arguments{
\item{model}{a model defined as a compound [list]}
}
\value{
a [list] with the most recent states
}
\description{
Save the state variables in a vector
}
