% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-interface.R
\name{init_states_L}
\alias{init_states_L}
\title{Initialize the Lstates}
\usage{
init_states_L(model)
}
\arguments{
\item{model}{a model defined as a compound [list]}
}
\value{
the model, a compound [list]
}
\description{
Initialize the Lstates
}
