% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-interface.R
\name{init_states_M}
\alias{init_states_M}
\title{Save state variables}
\usage{
init_states_M(model)
}
\arguments{
\item{model}{a model defined as a compound [list]}
}
\value{
the model, a compound [list]
}
\description{
Save state variables
}
