% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_point_sets.R
\name{lattice}
\alias{lattice}
\title{Set up a lattice with n x n points}
\usage{
lattice(n, mn, mx)
}
\arguments{
\item{n}{the number of points}

\item{mn}{the lower limit for location values}

\item{mx}{the upper limit for location values}
}
\value{
a set of xy locations
}
\description{
Set up a lattice with n x n points
}
