% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_graphs.R
\name{make_all_graphs.BQ}
\alias{make_all_graphs.BQ}
\title{Make graphs for a BQ model}
\usage{
\method{make_all_graphs}{BQ}(model)
}
\arguments{
\item{model}{a ramp.micro model object}
}
\value{
a ramp.micro model object
}
\description{
Make graphs for all the KGV objects, bigM, and bigMM
}
