% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{make_all_graphs.BQS}
\alias{make_all_graphs.BQS}
\title{Make graphs for a BQS model}
\usage{
\method{make_all_graphs}{BQS}(model)
}
\arguments{
\item{model}{a ramp.micro model object}
}
\value{
a ramp.micro model object
}
\description{
Make graphs for all the KGV objects and bigM
}
