% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_graphs.R
\name{make_all_graphs}
\alias{make_all_graphs}
\title{Make graphs}
\usage{
make_all_graphs(model)
}
\arguments{
\item{model}{a ramp.micro model object}
}
\value{
a ramp.micro model object
}
\description{
Make graphs for all the KGV objects and bigM
}
