% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{make_all_graphs_common}
\alias{make_all_graphs_common}
\title{Make all of the graphs}
\usage{
make_all_graphs_common(model)
}
\arguments{
\item{model}{a ramp.micro model object}
}
\value{
a ramp.micro model object
}
\description{
Make common graphs for all the KGV objects and bigM
}
