% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_graphs.R
\name{make_common_graphs}
\alias{make_common_graphs}
\title{Make all of the graphs}
\usage{
make_common_graphs(model)
}
\arguments{
\item{model}{a ramp.micro model object}
}
\value{
a ramp.micro model object
}
\description{
Make common graphs for all the KGV objects and bigM
}
