% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convex_hulls.R
\name{make_convex_hull_xy}
\alias{make_convex_hull_xy}
\title{Make the convex hull for the i^th community}
\usage{
make_convex_hull_xy(xy)
}
\arguments{
\item{xy}{the point set}
}
\value{
the xy points that define a convex hull

a [list] with a hull
}
\description{
Make the convex hull for the i^th community
}
