% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_convex_hulls.R
\name{make_convex_hulls}
\alias{make_convex_hulls}
\title{Make a set of convex hulls to visualize the community}
\usage{
make_convex_hulls(model, net, cut = NULL, f_color = viridis::turbo)
}
\arguments{
\item{model}{a model defined as a compound [list]}

\item{net}{a network object}

\item{cut}{if !null, the number of communities for igraph::cut_at}

\item{f_color}{a function that returns a list of colors (e.g. viridis::turbo)}
}
\value{
the network object with convex hulls attached
}
\description{
Make a set of convex hulls to visualize the community
}
