% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-BQ.R
\name{make_demography_BQ}
\alias{make_demography_BQ}
\title{Make the demographic matrices for the BQ model}
\usage{
make_demography_BQ(model)
}
\arguments{
\item{model}{a model defined as a compound [list]}
}
\value{
the model

model, a compound [list]
}
\description{
Make the demographic matrices for the BQ model
}
