% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-BQS.R
\name{make_demography_BQS}
\alias{make_demography_BQS}
\title{Make the demographic matrices for the BQS model}
\usage{
make_demography_BQS(model)
}
\arguments{
\item{model}{a model defined as a compound [list]}
}
\value{
model, a compound [list]
}
\description{
Make the demographic matrices for the BQS model
}
