% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{make_graph_obj}
\alias{make_graph_obj}
\title{Make graph object}
\usage{
make_graph_obj(M, type = "b", tag = "")
}
\arguments{
\item{M}{a matrix}

\item{type}{"b" for blood feeding; "q" for egg laying}

\item{tag}{a string for plotting}
}
\value{
a graph object
}
\description{
From a matrix, M, return a graph object
the clusters from the walktrap algorithm &
for the symmetric graph defined by M + t(M)
the clusters for greedy_clusters
}
