% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/K_matrices_make.R
\name{make_Kbb}
\alias{make_Kbb}
\title{Compute dispersal matrix to lay eggs within one feeding cycle}
\usage{
make_Kbb(model)
}
\arguments{
\item{model}{a model defined as a compound [list]}
}
\value{
the model, a compound [list]
}
\description{
Compute dispersal matrix to lay eggs within one feeding cycle
}
