% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/K_matrices.R
\name{make_Kbq.BQ}
\alias{make_Kbq.BQ}
\title{Compute net dispersal matrix to blood feed within one feeding cycle}
\usage{
\method{make_Kbq}{BQ}(model, Tmax = 100)
}
\arguments{
\item{model}{a model defined as a compound [list]}

\item{Tmax}{the number of time steps}
}
\value{
the model, a compound [list]
}
\description{
Compute net dispersal matrix to blood feed within one feeding cycle
}
