% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/K_matrices_make.R
\name{make_Kbq}
\alias{make_Kbq}
\title{Compute dispersal matrix to blood feed within one feeding cycle}
\usage{
make_Kbq(model, Tmax)
}
\arguments{
\item{model}{a model defined as a compound [list]}

\item{Tmax}{the number of time steps}
}
\value{
the model, a compound [list]
}
\description{
Compute dispersal matrix to blood feed within one feeding cycle
}
