% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_model.R
\name{make_model_unif}
\alias{make_model_unif}
\title{Setup a model on a random uniform set of points}
\usage{
make_model_unif(N, kFb, kFq, q_outside = TRUE)
}
\arguments{
\item{N}{the size of the big lattice}

\item{kFb}{a blood feeding dispersal kernel}

\item{kFq}{an egg laying dispersal kernel}

\item{q_outside}{a [logical] switch: if true, habitats define the big lattice}
}
\value{
a model as a compound [list]
}
\description{
Setup a model on a random uniform set of points
}
