% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movie.R
\name{make_movie.BQ}
\alias{make_movie.BQ}
\title{Make a movie of a simulation with a BQ model}
\usage{
\method{make_movie}{BQ}(model, moviestemx)
}
\arguments{
\item{model}{a model, defined as a [list]}

\item{moviestemx}{the stem name}
}
\value{
invisible(NULL)
}
\description{
Make a movie of a simulation with a BQ model
}
