% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_movie.R
\name{make_movie.BQS}
\alias{make_movie.BQS}
\title{Make a movie of a simulation with a BQS model}
\usage{
\method{make_movie}{BQS}(model, moviestemx = "BQSmovie")
}
\arguments{
\item{model}{a model, defined as a [list]}

\item{moviestemx}{the stem name}
}
\value{
invisible(NULL)
}
\description{
Make a movie of a simulation with a BQS model
}
