% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movie.R
\name{make_movie}
\alias{make_movie}
\title{Make a movie}
\usage{
make_movie(model, moviestemx)
}
\arguments{
\item{model}{a model, defined as a [list]}

\item{moviestemx}{the stem name}
}
\value{
invisible(NULL)
}
\description{
Make a movie
}
