% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-BQS.R
\name{make_Psi_BQS}
\alias{make_Psi_BQS}
\title{Make dispersal matrices for the BQS model}
\usage{
make_Psi_BQS(model)
}
\arguments{
\item{model}{a model defined as a compound [list]}
}
\value{
model, a compound [list]
}
\description{
Make dispersal matrices for the BQS model
}
