% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_tiles.R
\name{make_tiles.BQS}
\alias{make_tiles.BQS}
\title{Make a set of tiles that tesselate space for the BQS model}
\usage{
\method{make_tiles}{BQS}(model)
}
\arguments{
\item{model}{a model, defined as a compound [list]}
}
\value{
the model, a compound [list]
}
\description{
Make a set of tiles that tesselate space for the BQS model
}
