% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_convex_hulls.R
\name{mod_convex_hulls.i}
\alias{mod_convex_hulls.i}
\title{Make a set of convex hulls to visualize the community}
\usage{
mod_convex_hulls.i(
  model,
  i,
  cut = NULL,
  f_color = turbo,
  stretch = 0.1,
  lwd = 2
)
}
\arguments{
\item{model}{a model defined as a compound [list]}

\item{cut}{if !null, the number of communities for igraph::cut_at}

\item{f_color}{a function that returns a list of colors (e.g. viridis::turbo)}

\item{net}{a network object}
}
\value{
the network object with convex hulls attached
}
\description{
Make a set of convex hulls to visualize the community
}
