% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metapopulation.R
\name{net2meta}
\alias{net2meta}
\title{Use communities to define patches}
\usage{
net2meta(model, i, cut = NULL)
}
\arguments{
\item{model}{a ramp.micro model object}

\item{i}{which graph (see get_net)}

\item{cut}{cutat argument, # communities (optional)}
}
\value{
a ramp.micro model object
}
\description{
Use communities to define patches
}
