% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_points.R, R/viz_points.R
\name{outline_points_b}
\alias{outline_points_b}
\title{Add blood feeding points to a frame}
\usage{
outline_points_b(b, wts_b = 1, cx_b = 1, clr_b = "black", pw = 1)

outline_points_b(b, wts_b = 1, cx_b = 1, clr_b = "black", pw = 1)
}
\arguments{
\item{b}{a point set defining blood feeding sites}

\item{wts_b}{values to scale the point size}

\item{cx_b}{set the maximum cex for points}

\item{clr_b}{the outline color}

\item{pw}{power relationship for scaling point size: pw=1 is linear}
}
\value{
invisible(NULL)

invisible(NULL)
}
\description{
Add blood feeding points to a frame

Add blood feeding points to a frame
}
