% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_points.R
\name{outline_points_q}
\alias{outline_points_q}
\title{Add habitats to a frame}
\usage{
outline_points_q(q, wts_q = 1, cx_q = 1, clr_q = "black", pw = 1)
}
\arguments{
\item{q}{a point set defining egg laying sites}

\item{wts_q}{values to scale the point size}

\item{cx_q}{set the maximum cex for points}

\item{clr_q}{the color}

\item{pw}{power relationship for scaling point size: pw=1 is linear}
}
\value{
invisible(NULL)
}
\description{
Add habitats to a frame
}
