% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_points.R
\name{outline_points_s}
\alias{outline_points_s}
\title{Add habitats to a frame}
\usage{
outline_points_s(s, wts_s = 1, cx_s = 1, clr_s = "black", pw = 1)
}
\arguments{
\item{s}{a point set defining egg laying sites}

\item{wts_s}{values to scale the point size}

\item{cx_s}{set the maximum cex for points}

\item{clr_s}{the color}

\item{pw}{power relationship for scaling point size: pw=1 is linear}
}
\value{
invisible(NULL)
}
\description{
Add habitats to a frame
}
