% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_Psi.R
\name{plot_all_Psi.BQ}
\alias{plot_all_Psi.BQ}
\title{Visualize the one-bout dispersal matrices for a BQ model}
\usage{
\method{plot_all_Psi}{BQ}(
  model,
  cx_D = 2,
  cx_S = 0.3,
  min_edge_frac = 0.01,
  r = 0.01,
  arw_lng = 0.05,
  lwd = 2
)
}
\arguments{
\item{model}{a model defined as a compound [list]}

\item{cx_D}{the maximum cex for the destination}

\item{cx_S}{the maximum cex for the source}

\item{min_edge_frac}{the fraction of the mass to plot}

\item{r}{the radius of a ring around destination points}

\item{arw_lng}{the arrow length}

\item{lwd}{scale the line width}
}
\value{
the model, a compound [list]
}
\description{
Visualize the one-bout dispersal matrices for a BQ model
}
