% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_convex_hulls.R
\name{plot_convex_hulls}
\alias{plot_convex_hulls}
\title{Add the convex hulls to a framed plot}
\usage{
plot_convex_hulls(net, stretch = 1.1, lwd = 2)
}
\arguments{
\item{net}{a network object}

\item{stretch}{make the hull slightly larger or slightly smaller}

\item{lwd}{hull line width}
}
\value{
invisible(NULL)
}
\description{
Add the convex hulls to a framed plot
}
