% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_G.R
\name{plot_dispersal_G}
\alias{plot_dispersal_G}
\title{Plot lifetime egg dispersal, per mosquito}
\usage{
plot_dispersal_G(
  model,
  cx_b = 0.3,
  cx_q = 2,
  min_edge_frac = 0.01,
  r = 0.01,
  arw_lng = 0.05,
  lwd = 2,
  lamp = 1,
  seg_clr = "lightblue",
  arw_clr = "salmon"
)
}
\arguments{
\item{model}{a model defined as a compound [list]}

\item{cx_b}{the maximum cex for blood feeding sites}

\item{cx_q}{the maximum cex for egg laying sites}

\item{min_edge_frac}{the fraction of the mass to plot}

\item{r}{the radius of a ring around destination points}

\item{arw_lng}{the arrow length}

\item{lwd}{scale the line width}

\item{lamp}{arrow width scaling factor}

\item{seg_clr}{the color to draw the segment (symmetric part)}

\item{arw_clr}{the color to draw the arrow (asymetric part)}
}
\value{
invisible(NULL)
}
\description{
Plot lifetime egg dispersal, per mosquito
}
