% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_K.R
\name{plot_Kbb}
\alias{plot_Kbb}
\title{Plot the matrix \eqn{K_{b \leftarrow b}}}
\usage{
plot_Kbb(
  model,
  cx_b = 2,
  cx_q = 0.3,
  min_edge_frac = 0.01,
  r = 0.02,
  arw_lng = 0.002,
  lwd = 2,
  arw_clr = "#e2739655",
  seg_clr = "#00000022",
  clr_b = "#cc444bCC",
  clr_q = "#4361eeCC"
)
}
\arguments{
\item{model}{a model defined as a compound [list]}

\item{cx_b}{set the maximum cex for b points}

\item{cx_q}{set the maximum cex for q points}

\item{min_edge_frac}{the fraction of the mass to plot}

\item{r}{the radius of a ring around destination points}

\item{arw_lng}{the arrow length}

\item{lwd}{scale the line width}

\item{arw_clr}{the color to draw the arrow (asymetric part)}

\item{seg_clr}{the color to draw the segment (symmetric part)}

\item{clr_b}{color for blood feeding sites}

\item{clr_q}{color for egg laying sites}
}
\description{
Plot the matrix \eqn{K_{b \leftarrow b}}
}
