% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_K.R
\name{plot_Kbq}
\alias{plot_Kbq}
\title{Plot the matrix \eqn{K_{b \leftarrow q}}: dispersal from \eqn{\left\{q\right\}} to \eqn{\left\{b\right\}}}
\usage{
plot_Kbq(
  model,
  cx_b = 2,
  cx_q = 0.3,
  min_edge_frac = 0.01,
  r = 0.02,
  arw_lng = 0.002,
  lwd = 2,
  clr_K = "#fe5f55CC",
  clr_b = "darkred",
  clr_q = "#4361eeCC"
)
}
\arguments{
\item{model}{a model defined as a compound [list]}

\item{cx_b}{set the maximum cex for b points}

\item{cx_q}{set the maximum cex for q points}

\item{min_edge_frac}{the fraction of the mass to plot}

\item{r}{the radius of a ring around destination points}

\item{arw_lng}{the arrow length}

\item{lwd}{scale the line width}

\item{clr_K}{the color for Kqb arrows}

\item{clr_b}{the color for b points}

\item{clr_q}{the color for q points}
}
\value{
invisible(NULL)
}
\description{
Plot the matrix \eqn{K_{b \leftarrow q}}: dispersal from \eqn{\left\{q\right\}} to \eqn{\left\{b\right\}}
}
