% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_K.R
\name{plot_Kqb}
\alias{plot_Kqb}
\title{Plot the matrix \eqn{K_{q \leftarrow b}}
dispersal from \eqn{\left\{b\right\}} to \eqn{\left\{q\right\}}}
\usage{
plot_Kqb(
  model,
  cx_b = 0.3,
  cx_q = 2,
  min_edge_frac = 0.01,
  r = 0.02,
  arw_lng = 0.002,
  lwd = 2,
  clr_K = "#4361eeCC",
  clr_b = "red",
  clr_q = "darkblue"
)
}
\arguments{
\item{model}{a model defined as a compound [list]}

\item{cx_b}{set the maximum cex for b points}

\item{cx_q}{set the maximum cex for q points}

\item{min_edge_frac}{the fraction of the mass to plot}

\item{r}{the radius of a ring around destination points}

\item{arw_lng}{the arrow length}

\item{lwd}{scale the line width}

\item{clr_K}{the color for Kqb arrows}

\item{clr_b}{the color for b points}

\item{clr_q}{the color for q points}
}
\value{
invisible(NULL)
}
\description{
Plot the matrix \eqn{K_{q \leftarrow b}}
dispersal from \eqn{\left\{b\right\}} to \eqn{\left\{q\right\}}
}
