% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_K.R
\name{plot_Kqq}
\alias{plot_Kqq}
\title{Plot the matrix \eqn{K_{q \leftarrow q}}}
\usage{
plot_Kqq(
  model,
  cx_b = 0.03,
  cx_q = 2,
  min_edge_frac = 0.01,
  r = 0.02,
  arw_lng = 0.002,
  lwd = 2,
  arw_clr = "#4361eeCC",
  clr_qA = "#abc4ff55",
  seg_clr = "#00000022",
  clr_qB = "#858ae399",
  clr_b = "#cc444bCC"
)
}
\arguments{
\item{model}{a model defined as a compound [list]}

\item{cx_b}{set the maximum cex for b points}

\item{cx_q}{set the maximum cex for q points}

\item{min_edge_frac}{the minimum fraction to plot edge}

\item{r}{the radius of a ring around destination points}

\item{arw_lng}{the arrow length}

\item{lwd}{scale the line width}

\item{arw_clr}{the color to draw the arrow (asymetric part)}

\item{clr_qA}{color for egg laying sites}

\item{seg_clr}{the color to draw the segment (symmetric part)}

\item{clr_qB}{color for egg laying sites}

\item{clr_b}{color for blood feeding sites}
}
\value{
invisible(NULL)
}
\description{
Plot the matrix \eqn{K_{q \leftarrow q}}
}
