% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_matrix.R
\name{plot_matrix_xx}
\alias{plot_matrix_xx}
\title{Visualize a dispersal matrix from one point set to itself}
\usage{
plot_matrix_xx(
  xy,
  M,
  mx_pt_sz = 2,
  pt_clr = "tomato",
  min_edge_frac = 0.01,
  r = 0.01,
  arw_lng = 0.05,
  lwd = 1,
  lamp = 1,
  arw_clr = "#e2739655",
  seg_clr = "#00000022",
  mtl = ""
)
}
\arguments{
\item{xy}{the point set}

\item{M}{the dispersal matrix}

\item{mx_pt_sz}{set cex for the largest point}

\item{pt_clr}{the point color}

\item{min_edge_frac}{the minimum fraction to plot edge}

\item{r}{the radius of a ring around destination points}

\item{arw_lng}{the arrow length}

\item{lwd}{set the maximum line width}

\item{lamp}{arrow width scaling factor}

\item{arw_clr}{the color to draw the arrow (asymetric part)}

\item{seg_clr}{the color to draw the segment (symmetric part)}

\item{mtl}{= plot title}
}
\value{
no visible return value
}
\description{
Visualize a dispersal matrix from one point set to itself
}
