% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_matrix.R
\name{plot_matrix_xy}
\alias{plot_matrix_xy}
\title{Visualize a dispersal matrix among two point sets}
\usage{
plot_matrix_xy(
  xy_launch,
  xy_land,
  M,
  mx_pt_sz1 = 0.7,
  mx_pt_sz2 = 1.5,
  pt_clr1 = "salmon",
  pt_clr2 = "lightblue",
  min_edge_frac = 0.01,
  r = 0,
  arw_lng = 0.02,
  arw_lwd = 2,
  lamp = 1,
  arw_clr = "tomato",
  mtl = ""
)
}
\arguments{
\item{xy_launch}{origin point set}

\item{xy_land}{destination point set}

\item{M}{the dispersal matrix}

\item{mx_pt_sz1}{the maximum cex for the launch point sest}

\item{mx_pt_sz2}{the maximum cex for the landing point set}

\item{pt_clr1}{point color for launch}

\item{pt_clr2}{point color for landing}

\item{min_edge_frac}{the minimum fraction to plot edge}

\item{r}{the radius of a ring around destination points}

\item{arw_lng}{the arrow length}

\item{arw_lwd}{scale the line width}

\item{lamp}{arrow width scaling factor}

\item{arw_clr}{the arrow color}

\item{mtl}{= plot title}
}
\value{
no visible return value
}
\description{
Visualize a dispersal matrix among two point sets
}
