% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metapopulation.R
\name{plot_meta}
\alias{plot_meta}
\title{Visualize the metapopulation approximation}
\usage{
plot_meta(model, i, cut = NULL, lwd = 2, bbend = 3, mtl = NULL)
}
\arguments{
\item{model}{a ramp.micro model object}

\item{i}{the graph}

\item{cut}{a cutat argument, the # of communities}

\item{lwd}{a plotting argument}

\item{bbend}{the bend in the arrows}

\item{mtl}{a title for the graph}
}
\value{
invisible(NULL)
}
\description{
Visualize the metapopulation approximation
}
