% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patches.R
\name{plot_patches}
\alias{plot_patches}
\title{Visualize the patchespopulation approximation}
\usage{
plot_patches(
  model,
  i,
  cut = NULL,
  f_color = viridis::turbo,
  stretch = 0.1,
  lwd = 2,
  bbend = 3,
  mtl = NULL
)
}
\arguments{
\item{model}{a ramp.micro model object}

\item{i}{the graph}

\item{cut}{a cutat argument, the # of communities}

\item{f_color}{a function that returns a list of colors (e.g. viridis::turbo)}

\item{stretch}{make the hull slightly larger or slightly smaller}

\item{lwd}{a plotting argument}

\item{bbend}{the bend in the arrows}

\item{mtl}{a title for the graph}
}
\value{
a ramp.micro model object
}
\description{
Visualize the patchespopulation approximation
}
