% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_points.R
\name{plot_points.BQS}
\alias{plot_points.BQS}
\title{Plot point sets for a model for a BQS model}
\usage{
\method{plot_points}{BQS}(
  model,
  bwts = 1,
  qwts = 1,
  swts = 1,
  pw = 1,
  max_pt_sz = 1,
  mtl = ""
)
}
\arguments{
\item{model}{a model defined as a compound [list]}

\item{bwts}{values to scale the point size for points in b}

\item{qwts}{values to scale the point size for points in q}

\item{swts}{values to scale the point size for points in s}

\item{pw}{power relationship for scaling point size: pw=1 is linear}

\item{max_pt_sz}{set the maximum cex for points}

\item{mtl}{a plot title}
}
\value{
no visible return value

invisible(NULL)
}
\description{
Plot point sets for a model for a BQS model
}
