% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_points.R, R/viz_points.R
\name{plot_points}
\alias{plot_points}
\title{Plot point sets for a model}
\usage{
plot_points(
  model,
  wts_b = 1,
  wts_q = 1,
  wts_s = 1,
  cx_b = 1,
  cx_q = 1,
  cx_s = 1,
  clr_b = "#fe5f55CC",
  clr_q = "skyblue",
  clr_s = "olivedrab2",
  pw = 1,
  mtl = ""
)

plot_points(
  model,
  wts_b = 1,
  wts_q = 1,
  wts_s = 1,
  cx_b = 1,
  cx_q = 1,
  cx_s = 1,
  clr_b = "#fe5f55CC",
  clr_q = "skyblue",
  clr_s = "olivedrab2",
  pw = 1,
  mtl = ""
)
}
\arguments{
\item{model}{a compound [list] defining a model}

\item{wts_b}{values to scale the point size for points in b}

\item{wts_q}{values to scale the point size for points in q}

\item{wts_s}{values to scale the point size for points in s}

\item{cx_b}{set the maximum cex for b points}

\item{cx_q}{set the maximum cex for q points}

\item{cx_s}{set the maximum cex for s points}

\item{clr_b}{the color for b points}

\item{clr_q}{the color for q points}

\item{clr_s}{the color for s points}

\item{pw}{power relationship for scaling point size: pw=1 is linear}

\item{mtl}{a plot title}
}
\value{
invisible(NULL)

invisible(NULL)
}
\description{
Plot point sets for a model

Plot point sets for a model
}
