% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_points.R
\name{plot_points_bqs}
\alias{plot_points_bqs}
\title{Plot point sets for a model for a BQS model}
\usage{
plot_points_bqs(
  b,
  q,
  s,
  bwts = 1,
  qwts = 1,
  swts = 1,
  pw = 1,
  max_pt_sz = 1,
  mtl = ""
)
}
\arguments{
\item{b}{a point set defining blood feeding sites}

\item{q}{a point set defining egg laying sites}

\item{s}{a point set defining sugar feeding sites}

\item{bwts}{values to scale the point size for points in b}

\item{qwts}{values to scale the point size for points in q}

\item{swts}{values to scale the point size for points in s}

\item{pw}{power relationship for scaling point size: pw=1 is linear}

\item{max_pt_sz}{set the maximum cex for points}

\item{mtl}{a plot title}
}
\value{
invisible(NULL)
}
\description{
Plot point sets for a model for a BQS model
}
