% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Psi_matrices.R
\name{plot_Psi.BQ}
\alias{plot_Psi.BQ}
\title{Visualize the one-bout dispersal matrices for a BQ model}
\usage{
\method{plot_Psi}{BQ}(
  model,
  max_pt_sz = 2,
  min_edge_frac = 0.01,
  r = 0.01,
  arw_lng = 0.05,
  lwd = 2
)

\method{plot_Psi}{BQ}(
  model,
  max_pt_sz = 2,
  min_edge_frac = 0.01,
  r = 0.01,
  arw_lng = 0.05,
  lwd = 2
)
}
\arguments{
\item{model}{a model defined as a compound [list]}

\item{max_pt_sz}{set the maximum cex for points}

\item{min_edge_frac}{the fraction of the mass to plot}

\item{r}{the radius of a ring around destination points}

\item{arw_lng}{the arrow length}

\item{lwd}{scale the line width}
}
\value{
the model, a compound [list]

the model, a compound [list]
}
\description{
Visualize the one-bout dispersal matrices for a BQ model

Visualize the one-bout dispersal matrices for a BQ model
}
