% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Psi_matrices.R
\name{plot_Psi_BQSmod}
\alias{plot_Psi_BQSmod}
\title{Visualize the one-bout dispersal matrices for a BQS model}
\usage{
plot_Psi_BQSmod(
  b,
  q,
  s,
  Psi_bb,
  Psi_qb,
  Psi_sb,
  Psi_bq,
  Psi_qq,
  Psi_sq,
  Psi_bs,
  Psi_qs,
  Psi_ss,
  max_pt_sz = 2,
  min_edge_frac = 0.01,
  r = 0.01,
  arw_lng = 0.05,
  lwd = 2
)
}
\arguments{
\item{b}{blood feeding sites point set}

\item{q}{egg laying sites point set}

\item{s}{sugar feeding sites point set}

\item{Psi_bb}{one bout dispersal matrix from b to b}

\item{Psi_qb}{one bout dispersal matrix from q to b}

\item{Psi_sb}{one bout dispersal matrix from s to b}

\item{Psi_bq}{one bout dispersal matrix from b to q}

\item{Psi_qq}{one bout dispersal matrix from q to q}

\item{Psi_sq}{one bout dispersal matrix from s to q}

\item{Psi_bs}{one bout dispersal matrix from b to s}

\item{Psi_qs}{one bout dispersal matrix from q to s}

\item{Psi_ss}{one bout dispersal matrix from s to s}

\item{max_pt_sz}{set the maximum cex for points}

\item{min_edge_frac}{the fraction of the mass to plot}

\item{r}{the radius of a ring around destination points}

\item{arw_lng}{the arrow length}

\item{lwd}{scale the line width}
}
\value{
no visible return value
}
\description{
Visualize the one-bout dispersal matrices for a BQS model
}
