% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_Psi.R
\name{plot_Psi_qq}
\alias{plot_Psi_qq}
\title{Visualize the one-bout dispersal matrices for a BQ model}
\usage{
plot_Psi_qq(
  b,
  q,
  Psi_qq,
  cx = 2,
  min_edge_frac = 0.01,
  r = 0.01,
  arw_lng = 0.05,
  lwd = 2
)
}
\arguments{
\item{b}{blood feeding sites point set}

\item{q}{egg laying sites point set}

\item{Psi_qq}{one bout dispersal matrix from q to q}

\item{cx}{the maximum cex}

\item{min_edge_frac}{the fraction of the mass to plot}

\item{r}{the radius of a ring around destination points}

\item{arw_lng}{the arrow length}

\item{lwd}{scale the line width}
}
\value{
no visible return value
}
\description{
Visualize the one-bout dispersal matrices for a BQ model
}
