% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{plot_subgraph}
\alias{plot_subgraph}
\title{Plot the ouptuts of a graph using}
\usage{
plot_subgraph(
  model,
  graph,
  cut = NULL,
  alg = "wt",
  f_color = viridis::turbo,
  min_edge_frac = 0.01,
  cx = 2,
  pw = 1,
  mtl = "",
  stretch = 0.1,
  r = 0.02,
  arw_lng = 0.05,
  lwd = 2
)
}
\arguments{
\item{model}{a `ramp.micro` model object}

\item{graph}{a graphswork object}

\item{cut}{optional arguent for cut_at}

\item{alg}{walktrap = "wt" or greedy = "gr"}

\item{f_color}{a function that returns a list of colors (e.g. viridis::turbo)}

\item{min_edge_frac}{the fraction of the mass to plot}

\item{cx}{cex for plotting points}

\item{pw}{power relationship for scaling point size: pw=1 is linear}

\item{mtl}{a plot title}

\item{stretch}{make the hull slightly larger or slightly smaller}

\item{r}{the radius of a ring around destination points}

\item{arw_lng}{the arrow length}

\item{lwd}{scale the line width}
}
\value{
invisible(NULL)
}
\description{
Plot the ouptuts of a graph using
}
