% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispersal_distances.R
\name{plotDDcmf}
\alias{plotDDcmf}
\title{Plot the CMF for distances dispersed}
\usage{
plotDDcmf(xMFobj, mtl = NULL, clr = "black", thresh = 0.99)
}
\arguments{
\item{xMFobj}{a [list] returned from [dispersalPMF]}

\item{mtl}{a title for the plot}

\item{clr}{the colors}

\item{thresh}{a threshold for}
}
\value{
invisible(NULL)
}
\description{
Plot the CMF for distances dispersed
}
