% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispersal_distances.R
\name{plotDDpmf}
\alias{plotDDpmf}
\title{Plot the PMF for distances dispersed}
\usage{
plotDDpmf(xMFobj, mtl = NULL, clr = "black", thresh = 0.99)
}
\arguments{
\item{xMFobj}{a [list] returned from [dispersalPMF]}

\item{mtl}{a title for the plot}

\item{clr}{the colors}

\item{thresh}{a threshold for}
}
\value{
invisible(NULL)
}
\description{
Plot the PMF for distances dispersed
}
