% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-basicL.R
\name{save_states_L.basicL}
\alias{save_states_L.basicL}
\title{Save state variables}
\usage{
\method{save_states_L}{basicL}(t, model)
}
\arguments{
\item{t}{the current time}

\item{model}{a model defined as a compound [list]}
}
\value{
the model, a compound [list]
}
\description{
Save state variables
}
