% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-interface.R
\name{save_states_L}
\alias{save_states_L}
\title{Save state variables}
\usage{
save_states_L(t, model)
}
\arguments{
\item{t}{the current time}

\item{model}{a model defined as a compound [list]}
}
\value{
the model, a compound [list]
}
\description{
Save state variables
}
