% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-BQS.R
\name{save_states_M.BQS}
\alias{save_states_M.BQS}
\title{Save the state variables in a vector}
\usage{
\method{save_states_M}{BQS}(t, model)
}
\arguments{
\item{t}{the current time}

\item{model}{a model defined as a compound [list]}
}
\value{
a [list] with the most recent states
}
\description{
Save the state variables in a vector
}
