% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-interface.R
\name{setup_adult_model}
\alias{setup_adult_model}
\title{Setup Adult Mosquito Module}
\usage{
setup_adult_model(
  model,
  b,
  q,
  s,
  dispersal_opts = list(),
  bionomic_opts = list(),
  eip = 15
)
}
\arguments{
\item{model}{a model defined as a compound [list]}

\item{b}{a point set defining blood feeding sites}

\item{q}{a point set defining egg laying sites}

\item{s}{a point set defining sugar feeding sites}

\item{dispersal_opts}{a [list] to overwrite defaults}

\item{bionomic_opts}{a [list] to overwrite defaults}

\item{eip}{the extrinsic incubation period}
}
\value{
a [list] defining an adult model
}
\description{
Setup Adult Mosquito Module
}
