% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-basicL.R
\name{setup_aquatic_model.basicL}
\alias{setup_aquatic_model.basicL}
\title{Setup an aquatic model}
\usage{
\method{setup_aquatic_model}{basicL}(model, aquatic_opts = list())
}
\arguments{
\item{model}{a model defined as a compound [list]}

\item{aquatic_opts}{a [list] to overwrite defaults}
}
\value{
a [list] defining an adult model
}
\description{
Setup an aquatic model
}
