% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-interface.R
\name{setup_aquatic_model}
\alias{setup_aquatic_model}
\title{Setup an aquatic model}
\usage{
setup_aquatic_model(model, aquatic_opts = list())
}
\arguments{
\item{model}{a model defined as a compound [list]}

\item{aquatic_opts}{a [list] to overwrite defaults}
}
\value{
a [list] defining an adult model
}
\description{
Setup an aquatic model
}
