% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-basicL.R
\name{setup_aquatic_model_basicL}
\alias{setup_aquatic_model_basicL}
\title{Set up the aquatic population model of class basicL}
\usage{
setup_aquatic_model_basicL(
  model,
  opts = list(),
  pL = 0.9,
  zeta = 0.01,
  theta = 0.9,
  xi = 0
)
}
\arguments{
\item{model}{a model defined as a compound [list]}

\item{opts}{a [list] of values that overwrite the defaults}

\item{pL}{base survival fraction}

\item{zeta}{reduced survival in response to mean crowding}

\item{theta}{base maturation fraction}

\item{xi}{delayed maturation in response to mean crowding}
}
\value{
the model, a compound [list]
}
\description{
Set up the aquatic population model of class basicL
}
