% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-BQ.R
\name{setup_bionomics_BQ}
\alias{setup_bionomics_BQ}
\title{Bionomic parameters set for the BQ model}
\usage{
setup_bionomics_BQ(
  model,
  opts = list(),
  pB = 0.96,
  pQ = 0.96,
  psiB = 0.9,
  psiQ = 0.9,
  ova = 20
)
}
\arguments{
\item{model}{a model defined as a compound [list]}

\item{opts}{a [list] of values that overwrite the defaults}

\item{pB}{the probability of surviving a blood feeding bout}

\item{pQ}{the probability of surviving an egg laying bout}

\item{psiB}{blood feeding success}

\item{psiQ}{egg laying success}

\item{ova}{female eggs laid during a successful bout, per female}
}
\value{
the parameters, a compound [list]
}
\description{
Bionomic parameters set for the BQ model
}
