% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-BQS.R
\name{setup_bionomics_BQS}
\alias{setup_bionomics_BQS}
\title{Bionomic parameters set for the BQ model}
\usage{
setup_bionomics_BQS(
  model,
  opts = list(),
  pB = 0.96,
  pQ = 0.96,
  pS = 0.96,
  psiB = 0.9,
  psiQ = 0.9,
  psiS = 0.9,
  sigb = 0.1,
  sigq = 0.1,
  sigf = 0.1,
  sigL = 0.1,
  ova = 20
)
}
\arguments{
\item{model}{a model defined as a compound [list]}

\item{opts}{a [list] of values that overwrite the defaults}

\item{pB}{the probability of surviving a blood feeding bout}

\item{pQ}{the probability of surviving an egg laying bout}

\item{pS}{the probability of surviving a sugar feeding bout}

\item{psiB}{blood feeding success}

\item{psiQ}{egg laying success}

\item{psiS}{sugar feeding success}

\item{sigb}{transition from blood to sugar feeding}

\item{sigq}{transition from egg laying to sugar feeding}

\item{sigf}{transition from egg laying to sugar feeding}

\item{sigL}{transition from emergence to sugar feeding}

\item{ova}{female eggs laid during a successful bout, per female}
}
\value{
the parameters, a compound [list]
}
\description{
Bionomic parameters set for the BQ model
}
