% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-BQS.R
\name{setup_model.BQS}
\alias{setup_model.BQS}
\title{Setup a BQS model for adult mosquitoes}
\usage{
setup_model.BQS(
  model,
  b,
  q,
  s,
  dispersal_opts = list(),
  bionomic_opts = list(),
  eip = 15
)
}
\arguments{
\item{model}{a model defined as a compound [list]}

\item{b}{a point set defining blood feeding sites}

\item{q}{a point set defining egg laying sites}

\item{s}{a point set defining sugar feeding sites}

\item{dispersal_opts}{a [list] to overwrite defaults}

\item{bionomic_opts}{a [list] to overwrite defaults}

\item{eip}{the extrinsic incubation period}
}
\value{
a [list] defining a BQ-class adult model
}
\description{
Setup a BQS model for adult mosquitoes
}
