% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_model.R
\name{setup_model}
\alias{setup_model}
\title{Setup a behavioral state, micro-simulation model}
\usage{
setup_model(
  b,
  q,
  s = c(),
  kFb = NULL,
  kFq = NULL,
  kFs = NULL,
  Mname = "BQ",
  Lname = "basicL",
  dispersal_opts = list(),
  bionomic_opts = list(),
  aquatic_opts = list(),
  M0_opts = list(),
  L0_opts = list(),
  eip = 15
)
}
\arguments{
\item{b}{a point set defining blood feeding sites}

\item{q}{a point set defining egg laying sites}

\item{s}{a point set defining sugar feeding sites, with a default null value}

\item{kFb}{a kernel shape for blood searching}

\item{kFq}{a kernal shape for aquatic habitat searching}

\item{kFs}{a kernel shape for sugar site searching}

\item{Mname}{the adult model name}

\item{Lname}{the aquatic model name}

\item{dispersal_opts}{a [list] to}

\item{bionomic_opts}{a [list] to overwrite defaults}

\item{aquatic_opts}{a [list] to overwrite defaults}

\item{M0_opts}{options to overwrite defaults}

\item{L0_opts}{options to overwrite defaults}

\item{eip}{the extrinsic incubation period}
}
\value{
a [list] defining a BQ-class adult model
}
\description{
Setup a behavioral state, micro-simulation model
}
