% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{SIM}
\alias{SIM}
\title{Simulate population dynamics}
\usage{
SIM(model, Tmax = 200)
}
\arguments{
\item{model}{a compound [list] defining a model}

\item{Tmax}{the last time step}
}
\value{
model
}
\description{
Simulate population dynamics
}
