% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convex_hulls.R
\name{stretch_convex_hull}
\alias{stretch_convex_hull}
\title{Stretch (or shrink) the convex hull for plotting}
\usage{
stretch_convex_hull(xy, fac)
}
\arguments{
\item{xy}{a convex hull}

\item{fac}{a factor to stretch (>1) or shrink (<1) the hull for plotting}
}
\value{
a new point set
}
\description{
Stretch (or shrink) the convex hull for plotting
}
