% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point_sets.R
\name{unif_xy}
\alias{unif_xy}
\title{Make a random xy point set}
\usage{
unif_xy(n, mn, mx)
}
\arguments{
\item{n}{the number of points}

\item{mn}{the lower limit for location values}

\item{mx}{the upper limit for location values}
}
\value{
a set of xy locations
}
\description{
Make a random xy point set
}
