% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_arrows.R
\name{add_arrows_xx}
\alias{add_arrows_xx}
\title{Draw arrows from a point set to itself}
\usage{
add_arrows_xx(
  xy,
  M,
  min_edge_frac = 0.99,
  r = 0,
  arw_lng = 0.1,
  lwd = 5,
  lamp = 1,
  arw_clr = "darkgreen",
  seg_clr = "#CCCCCC"
)
}
\arguments{
\item{xy}{the point set}

\item{M}{the dispersal matrix}

\item{min_edge_frac}{the fraction of the mass to plot}

\item{r}{the radius of a ring around destination points}

\item{arw_lng}{the arrow length}

\item{lwd}{scale the line width}

\item{lamp}{arrow width scaling factor}

\item{arw_clr}{the color to draw the arrow (asymetric part)}

\item{seg_clr}{the color to draw the segment (symmetric part)}
}
\value{
invisible(NULL)
}
\description{
Draw arrows from a point set to itself
}
