% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_arrows.R, R/viz_arrows.R
\name{add_arrows_xy}
\alias{add_arrows_xy}
\title{Draw arrows from one point set to another}
\usage{
add_arrows_xy(
  xy_launch,
  xy_land,
  M,
  min_edge_frac = 0.01,
  r = 0,
  arw_lng = 0.1,
  lwd = 2,
  lamp = 1,
  clr = "darkgreen"
)

add_arrows_xy(
  xy_launch,
  xy_land,
  M,
  min_edge_frac = 0.01,
  r = 0,
  arw_lng = 0.1,
  lwd = 2,
  lamp = 1,
  clr = "darkgreen"
)
}
\arguments{
\item{xy_launch}{origin point set}

\item{xy_land}{destination point set}

\item{M}{the dispersal matrix}

\item{min_edge_frac}{the fraction of mass to plot}

\item{r}{the radius of a ring around destination points}

\item{arw_lng}{the arrow length}

\item{lwd}{scale the line width}

\item{lamp}{arrow width scaling factor}

\item{clr}{the arrow color}
}
\value{
invisible(NULL)

invisible(NULL)
}
\description{
Draw arrows from one point set to another

Draw arrows from one point set to another
}
