% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_bent_arrows.R
\name{add_bent_arrows_xy}
\alias{add_bent_arrows_xy}
\title{Draw bent arrows from one point set to another}
\usage{
add_bent_arrows_xy(
  xy_launch,
  xy_land,
  M,
  mnwd = 0.05,
  bbend = 0,
  endd = 0.75,
  adj = 2,
  clr = "red"
)
}
\arguments{
\item{xy_launch}{origin point set}

\item{xy_land}{destination point set}

\item{M}{the dispersal matrix}

\item{mnwd}{the minimum width to plot}

\item{bbend}{a parameter to bend arrows}

\item{endd}{relative position of arrowhead}

\item{adj}{set the maximum cex for points}

\item{clr}{the color}
}
\value{
invisible(NULL)
}
\description{
Draw bent arrows from one point set to another
}
