% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_points.R
\name{add_points_b}
\alias{add_points_b}
\title{Add blood feeding points to a frame}
\usage{
add_points_b(b, wts_b = 1, cx_b = 1, clr_b = "#fe5f55CC", pw = 1)
}
\arguments{
\item{b}{a point set defining blood feeding sites}

\item{wts_b}{values to scale the point size}

\item{cx_b}{set the maximum cex for points}

\item{clr_b}{the color for b points}

\item{pw}{power relationship for scaling point size: pw=1 is linear}
}
\value{
invisible(NULL)
}
\description{
Add blood feeding points to a frame
}
