% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_points.R
\name{add_points_bb}
\alias{add_points_bb}
\title{Add blood feeding sites with and without the self loop}
\usage{
add_points_bb(
  b,
  M,
  pw = 1,
  max_pt_sz = 2,
  colA = "#cc444b66",
  colB = "#cc444bCC"
)
}
\arguments{
\item{b}{a point set defining blood feeding sites}

\item{M}{a matrix describing movement to b}

\item{pw}{power relationship for scaling point size: pw=1 is linear}

\item{max_pt_sz}{set the maximum cex for points}

\item{colA}{color with the self loop}

\item{colB}{color without the self loop}
}
\value{
invisible(NULL)
}
\description{
Add blood feeding sites with and without the self loop
}
