% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_points.R
\name{add_points_qq}
\alias{add_points_qq}
\title{Add habitats with and without the self loop}
\usage{
add_points_qq(q, M, pw = 1, max_pt_sz = 2, colA = "skyblue", colB = "skyblue3")
}
\arguments{
\item{q}{a point set defining egg laying sites}

\item{M}{a matrix describing movement to q}

\item{pw}{power relationship for scaling point size: pw=1 is linear}

\item{max_pt_sz}{set the maximum cex for points}

\item{colA}{color with the self loop}

\item{colB}{color without the self loop}
}
\value{
invisible(NULL)
}
\description{
Add habitats with and without the self loop
}
