% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_points.R
\name{add_points_s}
\alias{add_points_s}
\title{Add sugar feeding sites to a frame}
\usage{
add_points_s(s, wts = 1, pw = 1, max_pt_sz = 1, clr = "olivedrab2")
}
\arguments{
\item{s}{a point set defining sugar feeding sites}

\item{wts}{values to scale the point size}

\item{pw}{power relationship for scaling point size: pw=1 is linear}

\item{max_pt_sz}{set the maximum cex for points}

\item{clr}{the color}
}
\value{
invisible(NULL)
}
\description{
Add sugar feeding sites to a frame
}
