% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_model.R
\name{basic_analysis}
\alias{basic_analysis}
\title{Basic analysis}
\usage{
basic_analysis(model, burn = 200, Tx = 50, tol = 0.001, Tmax = 200)
}
\arguments{
\item{model}{a model defined as a compound [list]}

\item{burn}{the burn argument for [steady_state]}

\item{Tx}{block size argument for [steady_state]}

\item{tol}{tolerance for [steady_state]}

\item{Tmax}{runtime for [makeKGV]}
}
\value{
model, a compound [list]
}
\description{
Basic analysis
}
