% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-basicL.R
\name{compute_diffs_L.basicL}
\alias{compute_diffs_L.basicL}
\title{The some of squared differences between two sets of variables}
\usage{
\method{compute_diffs_L}{basicL}(model)
}
\arguments{
\item{model}{a model}
}
\value{
a [numeric] value, the sum of squared differences
}
\description{
The some of squared differences between two sets of variables
}
