% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-BQ.R
\name{compute_diffs_M.BQ}
\alias{compute_diffs_M.BQ}
\title{The sum of squared differences between two sets of variables}
\usage{
\method{compute_diffs_M}{BQ}(Mvars1, Mvars2)
}
\arguments{
\item{Mvars1}{variables describing adult mosquitoes, set 1}

\item{Mvars2}{variables describing adult mosquitoes, set 2}
}
\value{
a [numeric] value, the sum of squared differences
}
\description{
The sum of squared differences between two sets of variables
}
