% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_EGG.R
\name{compute_G.BQ}
\alias{compute_G.BQ}
\title{Compute the dispersion of eggs after one feeding cycle for the BQ model}
\usage{
\method{compute_G}{BQ}(model, Tmax = 50)
}
\arguments{
\item{model}{a compound [list] defining a model}

\item{Tmax}{the last time step}
}
\value{
the model, a compound [list]
}
\description{
Compute the dispersion of eggs after one feeding cycle for the BQ model
}
