% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egg_laying.R
\name{compute_GG}
\alias{compute_GG}
\title{Compute the dispersion of eggs after one feeding cycle for the BQ model}
\usage{
compute_GG(model)
}
\arguments{
\item{model}{a compound [list] defining a model}
}
\value{
the model, a compound [list]
}
\description{
Compute the dispersion of eggs after one feeding cycle for the BQ model
}
