% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_K_matrices.R
\name{compute_Kbq.BQ}
\alias{compute_Kbq.BQ}
\title{Compute dispersal matrix to lay eggs within one feeding cycle}
\usage{
\method{compute_Kbq}{BQ}(model, Tmax = 100)
}
\arguments{
\item{model}{a model defined as a compound [list]}

\item{Tmax}{the number of time steps}
}
\value{
the model, a compound [list]
}
\description{
Compute dispersal matrix to lay eggs within one feeding cycle
}
