% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_K_matrices.R
\name{compute_Kqb.BQS}
\alias{compute_Kqb.BQS}
\title{Compute dispersal to lay eggs within one feeding cycle}
\usage{
\method{compute_Kqb}{BQS}(model, Tmax = 200)
}
\arguments{
\item{model}{a model defined as a compound [list]}

\item{Tmax}{the number of time steps}
}
\value{
the model, a compound [list]
}
\description{
Compute dispersal to lay eggs within one feeding cycle
}
