% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectorial_capacity.R
\name{compute_V.BQS}
\alias{compute_V.BQS}
\title{Compute the potential dispersion of parasites after one feeding cycle for the BQS model}
\usage{
\method{compute_V}{BQS}(model, Tmax = 100)
}
\arguments{
\item{model}{a compound [list] defining a model}

\item{Tmax}{the last time step}
}
\value{
the model, a compound [list]
}
\description{
Compute the potential dispersion of parasites after one feeding cycle for the BQS model
}
