% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispersal_distances.R
\name{dispersal_PMF}
\alias{dispersal_PMF}
\title{Compute distances dispersed}
\usage{
dispersal_PMF(xy, M)
}
\arguments{
\item{xy}{a set of point locations}

\item{M}{a dispersal matrix}
}
\value{
a [list] with the PMF, CMF, mean distance dispersed, and the distance matrix
}
\description{
Compute distances dispersed
}
