% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{make_kF_exp}
\alias{make_kF_exp}
\title{Make an exponential function to weight points by distance}
\usage{
make_kF_exp(k = 1, s = 2, gamma = 1)
}
\arguments{
\item{k}{decay by distance}

\item{s}{a scale parameter}

\item{gamma}{a shape parameter}
}
\value{
a function
}
\description{
Make an exponential function to weight points by distance
}
