% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_kernels.R
\name{make_kF_pwr}
\alias{make_kF_pwr}
\title{Make an power function to weight points by distance}
\usage{
make_kF_pwr(delta = 1, s = 1)
}
\arguments{
\item{delta}{the power on distance}

\item{s}{a shape function}
}
\value{
a function
}
\description{
Make an power function to weight points by distance
}
