% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/K_matrices_make.R
\name{make_Kqb.BQ}
\alias{make_Kqb.BQ}
\title{Compute dispersal matrix to lay eggs within one feeding cycle}
\usage{
\method{make_Kqb}{BQ}(model, Tmax = 100)
}
\arguments{
\item{model}{a model defined as a compound [list]}

\item{Tmax}{the number of time steps}
}
\value{
the model, a compound [list]
}
\description{
Compute dispersal matrix to lay eggs within one feeding cycle
}
