% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_Psi_matrices.R
\name{make_Psi_xx}
\alias{make_Psi_xx}
\title{Make Psi from a source point set to a destination point set}
\usage{
make_Psi_xx(S, kF = make_kF_exp(), w = 1, stay = 0)
}
\arguments{
\item{S}{a set of source points}

\item{kF}{a kernel weighting function}

\item{w}{linear weights for the destinations}

\item{stay}{the fraction that stays}
}
\value{
a [matrix], dispersal from S to S
}
\description{
Make Psi from a source point set to a destination point set
}
