% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_Psi_matrices.R
\name{make_Psi_xy}
\alias{make_Psi_xy}
\title{Make Psi from a source point set to a destination point set}
\usage{
make_Psi_xy(S, D, kF = make_kF_exp(), w = 1)
}
\arguments{
\item{S}{a set of source points}

\item{D}{a set of destination points}

\item{kF}{a kernel weighting function}

\item{w}{linear weights for the destinations}
}
\value{
a [matrix], dispersal from S to D
}
\description{
Make Psi from a source point set to a destination point set
}
