% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_convex_hulls.R
\name{mod_convex_hulls}
\alias{mod_convex_hulls}
\title{Make a set of convex hulls to visualize the community}
\usage{
mod_convex_hulls(
  model,
  net,
  cut = NULL,
  clrs = NULL,
  f_color = viridis::turbo,
  stretch = 0.1,
  lwd = 2
)
}
\arguments{
\item{model}{a model defined as a compound [list]}

\item{net}{a network object}

\item{cut}{if !null, the number of communities for igraph::cut_at}

\item{clrs}{a list of colors}

\item{f_color}{a function that returns a list of colors (e.g. viridis::turbo)}

\item{stretch}{a stretch factor for the hulls}

\item{lwd}{the line width for plotting}
}
\value{
the network object with convex hulls attached
}
\description{
Make a set of convex hulls to visualize the community
}
