% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patches.R
\name{net2patches}
\alias{net2patches}
\title{Use communities to define patches}
\usage{
net2patches(model, i, cut = NULL)
}
\arguments{
\item{model}{a ramp.micro model object}

\item{i}{which graph (see get_graph)}

\item{cut}{cutat argument, # communities (optional)}
}
\value{
a ramp.micro model object
}
\description{
Use communities to define patches
}
