% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Psi_matrices.R
\name{plot_all_Psi_BQ}
\alias{plot_all_Psi_BQ}
\title{Visualize the one-bout dispersal matrices for a BQ model}
\usage{
plot_all_Psi_BQ(
  b,
  q,
  Psi_bb,
  Psi_qb,
  Psi_bq,
  Psi_qq,
  max_pt_sz = 2,
  min_edge_frac = 0.01,
  r = 0.01,
  arw_lng = 0.05,
  lwd = 2
)
}
\arguments{
\item{b}{blood feeding sites point set}

\item{q}{egg laying sites point set}

\item{Psi_bb}{one bout dispersal matrix from b to b}

\item{Psi_qb}{one bout dispersal matrix from q to b}

\item{Psi_bq}{one bout dispersal matrix from b to q}

\item{Psi_qq}{one bout dispersal matrix from q to q}

\item{max_pt_sz}{set the maximum cex for points}

\item{min_edge_frac}{the fraction of the mass to plot}

\item{r}{the radius of a ring around destination points}

\item{arw_lng}{the arrow length}

\item{lwd}{scale the line width}
}
\value{
no visible return value
}
\description{
Visualize the one-bout dispersal matrices for a BQ model
}
