% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/K_matrices_plot.R
\name{plot_Kqb}
\alias{plot_Kqb}
\title{Plot the matrix Kqb: dispersal from {b} to {q}}
\usage{
plot_Kqb(
  model,
  max_pt_sz = 2,
  min_edge_frac = 0.01,
  r = 0.02,
  arw_lng = 0.002,
  lwd = 2,
  clr_K = "#4361eeCC",
  clr_b = "red",
  clr_q = "darkblue"
)
}
\arguments{
\item{model}{a model defined as a compound [list]}

\item{max_pt_sz}{set the maximum cex for points}

\item{min_edge_frac}{the fraction of the mass to plot}

\item{r}{the radius of a ring around destination points}

\item{arw_lng}{the arrow length}

\item{lwd}{scale the line width}

\item{clr_K}{the color for Kqb arrows}

\item{clr_b}{the color for b points}

\item{clr_q}{the color for q points}
}
\value{
invisible(NULL)
}
\description{
Plot the matrix Kqb: dispersal from {b} to {q}
}
