% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/K_matrices.R
\name{plot_Kqq}
\alias{plot_Kqq}
\title{Plot the matrix Kqq}
\usage{
plot_Kqq(
  model,
  max_pt_sz = 2,
  min_edge_frac = 0.01,
  r = 0.02,
  arw_lng = 0.002,
  lwd = 2,
  arw_clr = "#abc4ff55",
  seg_clr = "#00000022",
  clr_q = "#858ae399",
  clr_b = "#cc444bCC"
)
}
\arguments{
\item{model}{a model defined as a compound [list]}

\item{max_pt_sz}{set the maximum cex for points}

\item{min_edge_frac}{the minimum fraction to plot edge}

\item{r}{the radius of a ring around destination points}

\item{arw_lng}{the arrow length}

\item{lwd}{scale the line width}

\item{arw_clr}{the color to draw the arrow (asymetric part)}

\item{seg_clr}{the color to draw the segment (symmetric part)}

\item{clr_q}{color for egg laying sites}

\item{clr_b}{color for blood feeding sites}
}
\value{
invisible(NULL)
}
\description{
Plot the matrix Kqq
}
