% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_points.R
\name{plot_points_bq}
\alias{plot_points_bq}
\title{Plot point sets for a bq model}
\usage{
plot_points_bq(
  b,
  q,
  wts_b = 1,
  wts_q = 1,
  cx_b = 1,
  cx_q = 1,
  clr_b = "#fe5f55CC",
  clr_q = "skyblue",
  pw = 1,
  mtl = ""
)
}
\arguments{
\item{b}{a point set defining blood feeding sites}

\item{q}{a point set defining egg laying sites}

\item{wts_b}{values to scale the point size for points in b}

\item{wts_q}{values to scale the point size for points in q}

\item{cx_b}{set the maximum cex for b points}

\item{cx_q}{set the maximum cex for q points}

\item{clr_b}{the color for b points}

\item{clr_q}{the color for q points}

\item{pw}{power relationship for scaling point size: pw=1 is linear}

\item{mtl}{a plot title}
}
\value{
invisible(NULL)
}
\description{
Plot point sets for a bq model
}
