% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-interface.R
\name{save_states_M}
\alias{save_states_M}
\title{Store State Variables for an Adult Mosquito Model}
\usage{
save_states_M(t, model)
}
\arguments{
\item{t}{the current time}

\item{model}{a model defined as a compound [list]}
}
\value{
the model, a compound [list]
}
\description{
Store State Variables for an Adult Mosquito Model
}
